Completed working on the new stored procedure, modified the input values to the correct datatypes
Directly injected the objects in a way that would avoid SQL injection vulnerabilities and unexpected behavior

Values succefully add now

Next
Work on the remaining procedure logic that migrates values to the other tables

+++ Ms_DataCollector/dbo/Tables/Ms_Collection_II.sql
@@ -3,7 +3,7 @@
 		DocNum				INT NOT NULL IDENTITY (1, 1),
         DocEntry			VARCHAR(255) NOT NULL,
 		TranId			    VARCHAR(255) NULL,
-	    LongDate			    INT NOT NULL,
+	    LongDate			BIGINT NOT NULL,
         DocDateTime		    DATETIME NOT NULL,
 		Recepient		    VARCHAR(255) NOT NULL,
         AccNo	            varchar(255),
